Add password reset feature with emails
This completes the server portion of a password reset system. Tokens are
generated by mongoose-token using the node crypto library. The token is
stored in the DB and valid for 1 hour. An email containing the token is
sent to them using Mailgun. The free tier provides 300-400 free emails
per day.

The client portion of the feature should be a page that is accessible
without authentication containing a form for the token, the user email,
and the new password. This form should post to the reset_token endpoint from
this commit.